fix heisenbug with apollo-client caching
This fixes an issue with apollo-client's caching
and fetching system. Our current method of
extracting object IDs does not also include the
type of the object in its return value.

Consequently, apollo thinks that an object of
show ID 1 is also the entry id 1. The only
side-effect of this I have observed is that
the __typename of an embedded Entry is
occasionally 'Show', which is clearly incorrect.

This fixes the issue by prefixing an object's
ID with its typename.