Fix Regression in Ordering
The test case 'Gets all shows in order if requested' was not actually testing anything.
It was succeeding based on whatever random ordering MySQL returned.

Based on previous changes to the 'showQuery' resolver,
specifying an order without providing a studentUsername
would not honor the desired order. This restores that functionality.